Improve autocomplete module * Separated controllers. * Separated rest services. * Fixed issues.
